# docs/swagger.yaml
basePath: /
definitions:
...
  dto.APIResponse:
    properties:
      data:
        description: Dữ liệu trả về (nếu có)
      error:
        description: Lỗi (nếu có)
        example: ""
        type: string
      message:
        description: Thông báo
        example: Thao tác thành công
        type: string
      success:
        description: Request thành công hay thất bại
        example: true
        type: boolean
    type: object
  dto.ActiveSessionsResponse:
    properties:
      active_sessions:
        example: 3
        type: integer
    type: object
...
  dto.AuthResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      expires_in:
        description: seconds
        example: 900
        type: integer
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
...
  dto.LoginRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: john_doe
        type: string
    required:
    - password
    - username
    type: object
  dto.LogoutAllResponse:
    properties:
      message:
        example: Đã đăng xuất khỏi tất cả thiết bị
        type: string
      revoked_sessions:
        example: 3
        type: integer
    type: object
...
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    required:
    - refresh_token
    type: object
  dto.RefreshTokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      expires_in:
        description: seconds
        example: 900
        type: integer
      token_type:
        example: Bearer
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: password123
        maxLength: 100
        minLength: 6
        type: string
      username:
        example: john_doe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  dto.ResendCooldownResponse:
    properties:
      message:
        example: Vui lòng đợi trước khi gửi lại
        type: string
      remaining_seconds:
        example: 45
        type: integer
    type: object
  dto.ResendVerificationResponse:
    properties:
      message:
        example: Email xác thực đã được gửi lại
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      created_at:
        example: 24/01/2026 10:00
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: uuid-123
        type: string
      is_active:
        example: true
        type: boolean
      is_email_verified:
        example: true
        type: boolean
      role:
        example: staff
        type: string
      updated_at:
        example: 24/01/2026 10:30
        type: string
      username:
        example: john_doe
        type: string
    type: object
  dto.VerifyEmailRequest:
    properties:
      token:
        example: abc123def456...
        type: string
    required:
    - token
    type: object
  dto.VerifyEmailResponse:
    properties:
      message:
        example: Email đã được xác thực thành công
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@restaurant.local
    name: API Support
  description: API quản lý menu nhà hàng - Xây dựng theo Clean Architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Restaurant Menu API
  version: 2.0.0
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Đăng nhập và nhận tokens
      parameters:
      - description: Thông tin đăng nhập
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Đăng nhập
      tags:
      - Auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Đăng xuất và revoke access token hiện tại
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Đăng xuất
      tags:
      - Auth
  /api/auth/logout-all:
    post:
      consumes:
      - application/json
      description: Đăng xuất và revoke tất cả tokens của user (tất cả devices)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LogoutAllResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Đăng xuất tất cả thiết bị
      tags:
      - Auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Làm mới access token từ refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RefreshTokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Làm mới access token
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Đăng ký tài khoản mới (Customer)
      parameters:
      - description: Thông tin đăng ký
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Đăng ký tài khoản
      tags:
      - Auth
  /api/auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Gửi lại email xác thực cho user (cần đăng nhập)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ResendVerificationResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "429":
          description: Too Many Requests
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ResendCooldownResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Gửi lại email xác thực
      tags:
      - Auth
  /api/auth/sessions:
    get:
      description: Lấy số lượng thiết bị đang đăng nhập
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ActiveSessionsResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Lấy số session active
      tags:
      - Auth
  /api/auth/verify-email:
    post:
      consumes:
      - application/json
      description: Xác thực email của user với verification token
      parameters:
      - description: Verification token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.VerifyEmailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Xác thực email
      tags:
      - Auth
...
schemes:
- http
swagger: "2.0"